% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/things_an_architect_should_know.R
\docType{data}
\name{things_an_architect_should_know}
\alias{things_an_architect_should_know}
\title{"Two Hundred Fifty Things an Architect Should Know" as a data table.}
\format{
A data frame with 250 rows and 2 variables
\describe{
\item{n}{int, the order in which the thing was originally published.}
\item{thing}{character, a thing an architect should know}
}
}
\source{
\url{https://www.readingdesign.org/250-things}
}
\usage{
things_an_architect_should_know
}
\description{
A dataset containing all 250 things an architect should know, according to
\href{https://en.wikipedia.org/wiki/Michael_Sorkin}{Michael Sorkin}, an American
architectural and urban critic, designer, and educator. The list was originally
published in Sorkin's 2018 book, "What Goes Up: The Rights and Wrongs to a City".
}
\details{
Some things were modified to so that each thing made sense on its own.
These modifications are in parentheses.
}
\keyword{datasets}
