% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_should_an_architect_know.R
\name{what_should_an_architect_know}
\alias{what_should_an_architect_know}
\title{What an architect should know}
\usage{
what_should_an_architect_know(thing = sample(1:250, size = 1))
}
\arguments{
\item{thing}{a single numeric value or a vector of numeric values between 1 and 250.}
}
\value{
a random vector of things an architect should know
}
\description{
What an architect should know
}
\examples{

what_should_an_architect_know()

what_should_an_architect_know(1:10)

}
